+++ sdk/ai/azure-ai-inference/azure/ai/inference/_patch.py
@@ -16,6 +16,7 @@ Why do we patch auto-generated code?
    https://github.com/Azure/autorest.python/issues/2619 (all clients).
    Otherwise intellisense did not show the patched public methods on the client object,
    when the client is defined using context manager ("with" statement).
+6. Add support for load() method in ImageUrl class (see /models/_patch.py).
 
 """
 from __future__ import annotations
@@ -36,6 +37,7 @@ from typing import (
     overload,
     Type,
     TYPE_CHECKING,
+    Iterable,
     Iterator,
     TypeVar,
     Callable,
@@ -249,7 +251,7 @@ def llm_trace(func: Callable[_P, _R]) -> Callable[_P, _R]:
 
 def load_client(
     endpoint: str, credential: Union[AzureKeyCredential, "TokenCredential"], **kwargs: Any
-) -> Union[ChatCompletionsClientGenerated, EmbeddingsClientGenerated, ImageEmbeddingsClientGenerated]:
+) -> Union["ChatCompletionsClient", "EmbeddingsClient", "ImageEmbeddingsClient"]:
     """
     Load a client from a given endpoint URL. The method makes a REST API call to the `/info` route
     on the given endpoint, to determine the model type and therefore which client to instantiate.
@@ -264,7 +266,7 @@ def load_client(
      "2024-05-01-preview". Note that overriding this default value may result in unsupported
      behavior.
     :paramtype api_version: str
-    :return: The appropriate client associated with the given endpoint
+    :return: The appropriate synchronous client associated with the given endpoint
     :rtype: ~azure.ai.inference.ChatCompletionsClient or ~azure.ai.inference.EmbeddingsClient
      or ~azure.ai.inference.ImageEmbeddingsClient
     :raises ~azure.core.exceptions.HttpResponseError
@@ -284,7 +286,9 @@ def load_client(
     # TODO: Remove "completions" and "embedding" once Mistral Large and Cohere fixes their model type
     if model_info.model_type in (_models.ModelType.CHAT, "completion"):
         chat_completion_client = ChatCompletionsClient(endpoint, credential, **kwargs)
-        chat_completion_client._model_info = model_info  # pylint: disable=protected-access,attribute-defined-outside-init
+        chat_completion_client._model_info = ( # pylint: disable=protected-access,attribute-defined-outside-init
+            model_info
+        )
         return chat_completion_client
 
     if model_info.model_type in (_models.ModelType.EMBEDDINGS, "embedding"):
@@ -294,7 +298,9 @@ def load_client(
 
     if model_info.model_type == _models.ModelType.IMAGE_EMBEDDINGS:
         image_embedding_client = ImageEmbeddingsClient(endpoint, credential, **kwargs)
-        image_embedding_client._model_info = model_info  # pylint: disable=protected-access,attribute-defined-outside-init
+        image_embedding_client._model_info = (  # pylint: disable=protected-access,attribute-defined-outside-init
+            model_info
+        )
         return image_embedding_client
 
     raise ValueError(f"No client available to support AI model type `{model_info.model_type}`")
@@ -339,6 +345,7 @@ class ChatCompletionsClient(ChatCompletionsClientGenerated):
             Union[str, _models.ChatCompletionsToolSelectionPreset, _models.ChatCompletionsNamedToolSelection]
         ] = None,
         seed: Optional[int] = None,
+        model: Optional[str] = None,
         **kwargs: Any,
     ) -> _models.ChatCompletions: ...
 
@@ -362,8 +369,9 @@ class ChatCompletionsClient(ChatCompletionsClientGenerated):
             Union[str, _models.ChatCompletionsToolSelectionPreset, _models.ChatCompletionsNamedToolSelection]
         ] = None,
         seed: Optional[int] = None,
+        model: Optional[str] = None,
         **kwargs: Any,
-    ) -> _models.StreamingChatCompletions: ...
+    ) -> Iterable[_models.StreamingChatCompletionsUpdate]: ...
 
     @overload
     def complete(
@@ -385,8 +393,9 @@ class ChatCompletionsClient(ChatCompletionsClientGenerated):
             Union[str, _models.ChatCompletionsToolSelectionPreset, _models.ChatCompletionsNamedToolSelection]
         ] = None,
         seed: Optional[int] = None,
+        model: Optional[str] = None,
         **kwargs: Any,
-    ) -> Union[_models.StreamingChatCompletions, _models.ChatCompletions]:
+    ) -> Union[Iterable[_models.StreamingChatCompletionsUpdate], _models.ChatCompletions]:
         # pylint: disable=line-too-long
         """Gets chat completions for the provided chat messages.
         Completions support a wide variety of tasks and generate text that continues from or
@@ -468,10 +477,13 @@ class ChatCompletionsClient(ChatCompletionsClientGenerated):
          ~azure.ai.inference.models.ChatCompletionsNamedToolSelection
         :keyword seed: If specified, the system will make a best effort to sample deterministically
          such that repeated requests with the
-         same seed and parameters should return the same result. Determinism is not guaranteed.".
+         same seed and parameters should return the same result. Determinism is not guaranteed.
          Default value is None.
         :paramtype seed: int
-        :return: ChatCompletions for non-streaming, or StreamingChatCompletions for streaming.
+        :keyword model: ID of the specific AI model to use, if more than one model is available on the
+         endpoint. Default value is None.
+        :paramtype model: str
+        :return: ChatCompletions for non-streaming, or Iterable[StreamingChatCompletionsUpdate] for streaming.
         :rtype: ~azure.ai.inference.models.ChatCompletions or ~azure.ai.inference.models.StreamingChatCompletions
         :raises ~azure.core.exceptions.HttpResponseError
         """
@@ -483,7 +495,7 @@ class ChatCompletionsClient(ChatCompletionsClientGenerated):
         *,
         content_type: str = "application/json",
         **kwargs: Any,
-    ) -> Union[_models.StreamingChatCompletions, _models.ChatCompletions]:
+    ) -> Union[Iterable[_models.StreamingChatCompletionsUpdate], _models.ChatCompletions]:
         # pylint: disable=line-too-long
         """Gets chat completions for the provided chat messages.
         Completions support a wide variety of tasks and generate text that continues from or
@@ -495,7 +507,7 @@ class ChatCompletionsClient(ChatCompletionsClientGenerated):
         :keyword content_type: Body Parameter content-type. Content type parameter for JSON body.
          Default value is "application/json".
         :paramtype content_type: str
-        :return: ChatCompletions for non-streaming, or StreamingChatCompletions for streaming.
+        :return: ChatCompletions for non-streaming, or Iterable[StreamingChatCompletionsUpdate] for streaming.
         :rtype: ~azure.ai.inference.models.ChatCompletions or ~azure.ai.inference.models.StreamingChatCompletions
         :raises ~azure.core.exceptions.HttpResponseError
         """
@@ -507,7 +519,7 @@ class ChatCompletionsClient(ChatCompletionsClientGenerated):
         *,
         content_type: str = "application/json",
         **kwargs: Any,
-    ) -> Union[_models.StreamingChatCompletions, _models.ChatCompletions]:
+    ) -> Union[Iterable[_models.StreamingChatCompletionsUpdate], _models.ChatCompletions]:
         # pylint: disable=line-too-long
         # pylint: disable=too-many-locals
         """Gets chat completions for the provided chat messages.
@@ -519,7 +531,7 @@ class ChatCompletionsClient(ChatCompletionsClientGenerated):
         :keyword content_type: Body Parameter content-type. Content type parameter for binary body.
          Default value is "application/json".
         :paramtype content_type: str
-        :return: ChatCompletions for non-streaming, or StreamingChatCompletions for streaming.
+        :return: ChatCompletions for non-streaming, or Iterable[StreamingChatCompletionsUpdate] for streaming.
         :rtype: ~azure.ai.inference.models.ChatCompletions or ~azure.ai.inference.models.StreamingChatCompletions
         :raises ~azure.core.exceptions.HttpResponseError
         """
@@ -544,8 +556,9 @@ class ChatCompletionsClient(ChatCompletionsClientGenerated):
             Union[str, _models.ChatCompletionsToolSelectionPreset, _models.ChatCompletionsNamedToolSelection]
         ] = None,
         seed: Optional[int] = None,
+        model: Optional[str] = None,
         **kwargs: Any,
-    ) -> Union[_models.StreamingChatCompletions, _models.ChatCompletions]:
+    ) -> Union[Iterable[_models.StreamingChatCompletionsUpdate], _models.ChatCompletions]:
         # pylint: disable=line-too-long
         # pylint: disable=too-many-locals
         """Gets chat completions for the provided chat messages.
@@ -625,10 +638,13 @@ class ChatCompletionsClient(ChatCompletionsClientGenerated):
          ~azure.ai.inference.models.ChatCompletionsNamedToolSelection
         :keyword seed: If specified, the system will make a best effort to sample deterministically
          such that repeated requests with the
-         same seed and parameters should return the same result. Determinism is not guaranteed.".
+         same seed and parameters should return the same result. Determinism is not guaranteed.
          Default value is None.
         :paramtype seed: int
-        :return: ChatCompletions for non-streaming, or StreamingChatCompletions for streaming.
+        :keyword model: ID of the specific AI model to use, if more than one model is available on the
+         endpoint. Default value is None.
+        :paramtype model: str
+        :return: ChatCompletions for non-streaming, or Iterable[StreamingChatCompletionsUpdate] for streaming.
         :rtype: ~azure.ai.inference.models.ChatCompletions or ~azure.ai.inference.models.StreamingChatCompletions
         :raises ~azure.core.exceptions.HttpResponseError
         """
@@ -653,6 +669,7 @@ class ChatCompletionsClient(ChatCompletionsClientGenerated):
                 "frequency_penalty": frequency_penalty,
                 "max_tokens": max_tokens,
                 "messages": messages,
+                "model": model,
                 "presence_penalty": presence_penalty,
                 "response_format": response_format,
                 "seed": seed,
@@ -777,13 +794,6 @@ class EmbeddingsClient(EmbeddingsClientGenerated):
         :keyword content_type: Body Parameter content-type. Content type parameter for JSON body.
          Default value is "application/json".
         :paramtype content_type: str
-        :keyword extras: Extra parameters (in the form of string key-value pairs) that are not in the
-         standard request payload.
-         They will be passed to the service as-is in the root of the JSON request payload.
-         How the service handles these extra parameters depends on the value of the
-         ``extra-parameters``
-         HTTP request header. Default value is None.
-        :paramtype extras: dict[str, str]
         :keyword dimensions: Optional. The number of dimensions the resulting output embeddings should
          have.
          Passing null causes the model to use its default value.
@@ -1029,13 +1039,6 @@ class ImageEmbeddingsClient(ImageEmbeddingsClientGenerated):
         :keyword content_type: Body Parameter content-type. Content type parameter for JSON body.
          Default value is "application/json".
         :paramtype content_type: str
-        :keyword extras: Extra parameters (in the form of string key-value pairs) that are not in the
-         standard request payload.
-         They will be passed to the service as-is in the root of the JSON request payload.
-         How the service handles these extra parameters depends on the value of the
-         ``extra-parameters``
-         HTTP request header. Default value is None.
-        :paramtype extras: dict[str, str]
         :keyword dimensions: Optional. The number of dimensions the resulting output embeddings should
          have.
          Passing null causes the model to use its default value.