+++ sdk/ai/azure-ai-inference/samples/async_samples/sample_image_embeddings_async.py
@@ -11,11 +11,11 @@ USAGE:
     python sample_image_embeddings_async.py
 
     Set these two environment variables before running the sample:
-    1) IMAGE_EMBEDDINGS_ENDPOINT - Your endpoint URL, in the form 
+    1) AZURE_AI_IMAGE_EMBEDDINGS_ENDPOINT - Your endpoint URL, in the form 
         https://<your-deployment-name>.<your-azure-region>.inference.ai.azure.com
         where `your-deployment-name` is your unique AI Model deployment name, and
         `your-azure-region` is the Azure region where your model is deployed.
-    2) IMAGE_EMBEDDINGS_KEY - Your model key (a 32-character string). Keep it secret.
+    2) AZURE_AI_IMAGE_EMBEDDINGS_KEY - Your model key (a 32-character string). Keep it secret.
 """
 import asyncio
 
@@ -25,10 +25,10 @@ async def sample_image_embeddings_async():
     import base64
 
     try:
-        endpoint = os.environ["IMAGE_EMBEDDINGS_ENDPOINT"]
-        key = os.environ["IMAGE_EMBEDDINGS_KEY"]
+        endpoint = os.environ["AZURE_AI_IMAGE_EMBEDDINGS_ENDPOINT"]
+        key = os.environ["AZURE_AI_IMAGE_EMBEDDINGS_KEY"]
     except KeyError:
-        print("Missing environment variable 'IMAGE_EMBEDDINGS_ENDPOINT' or 'IMAGE_EMBEDDINGS_KEY'")
+        print("Missing environment variable 'AZURE_AI_IMAGE_EMBEDDINGS_ENDPOINT' or 'AZURE_AI_IMAGE_EMBEDDINGS_KEY'")
         print("Set them before running this sample.")
         exit()