+++ sdk/mobilenetwork/azure-mgmt-mobilenetwork/azure/mgmt/mobilenetwork/models/_mobile_network_management_client_enums.py
@@ -101,29 +101,29 @@ class InstallationReason(str, Enum, metaclass=CaseInsensitiveEnumMeta):
 
     NO_SLICES = "NoSlices"
     """The packet core has not been installed as the mobile network does not have any applicable
-    #: configured slices."""
+    configured slices."""
     NO_PACKET_CORE_DATA_PLANE = "NoPacketCoreDataPlane"
     """The packet core has not been installed as there is no configured data plane for this packet
-    #: core."""
+    core."""
     NO_ATTACHED_DATA_NETWORKS = "NoAttachedDataNetworks"
     """The packet core has not been installed as the packet core has no attached data networks."""
     PUBLIC_LAND_MOBILE_NETWORK_IDENTIFIER_HAS_CHANGED = "PublicLandMobileNetworkIdentifierHasChanged"
     """A reinstall is required as the packet core is running with out-of-date PLMN ID."""
     CONTROL_PLANE_ACCESS_INTERFACE_HAS_CHANGED = "ControlPlaneAccessInterfaceHasChanged"
     """A reinstall is required as the packet core is running with out-of-date control plane access
-    #: interface information."""
+    interface information."""
     USER_PLANE_ACCESS_INTERFACE_HAS_CHANGED = "UserPlaneAccessInterfaceHasChanged"
     """A reinstall is required as the packet core is running with out-of-date user plane core
-    #: interface."""
+    interface."""
     USER_PLANE_DATA_INTERFACE_HAS_CHANGED = "UserPlaneDataInterfaceHasChanged"
     """A reinstall is required as the packet core is running with out-of-date user plane access
-    #: interface."""
+    interface."""
     CONTROL_PLANE_ACCESS_VIRTUAL_IPV4_ADDRESSES_HAS_CHANGED = "ControlPlaneAccessVirtualIpv4AddressesHasChanged"
     """A reinstall is required as the packet core is running with out-of-date control plane access
-    #: network virtual IP address."""
+    network virtual IP address."""
     USER_PLANE_ACCESS_VIRTUAL_IPV4_ADDRESSES_HAS_CHANGED = "UserPlaneAccessVirtualIpv4AddressesHasChanged"
     """A reinstall is required as the packet core is running with out-of-date user plane access
-    #: network virtual IP address."""
+    network virtual IP address."""
 
 
 class InstallationState(str, Enum, metaclass=CaseInsensitiveEnumMeta):
@@ -165,6 +165,17 @@ class NaptEnabled(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """NAPT is disabled"""
 
 
+class NasEncryptionType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """The supported NAS Encryption types."""
+
+    NEA0_EEA0 = "NEA0/EEA0"
+    """NAS signaling is not encrypted."""
+    NEA1_EEA1 = "NEA1/EEA1"
+    """NAS signaling is encrypted with SNOW 3G cipher."""
+    NEA2_EEA2 = "NEA2/EEA2"
+    """NAS signaling is encrypted with AES cipher."""
+
+
 class ObsoleteVersion(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """Indicates whether this version is obsolete."""
 
@@ -201,10 +212,10 @@ class PlatformType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
 
     AKS_HCI = "AKS-HCI"
     """If this option is chosen, you must set one of "azureStackEdgeDevice", "connectedCluster" or
-    #: "customLocation". If multiple are set, they must be consistent with each other."""
+    "customLocation". If multiple are set, they must be consistent with each other."""
     THREE_P_AZURE_STACK_HCI = "3P-AZURE-STACK-HCI"
     """If this option is chosen, you must set one of "azureStackHciCluster", "connectedCluster" or
-    #: "customLocation". If multiple are set, they must be consistent with each other."""
+    "customLocation". If multiple are set, they must be consistent with each other."""
 
 
 class PreemptionCapability(str, Enum, metaclass=CaseInsensitiveEnumMeta):