+++ sdk/ai/azure-ai-inference/CHANGELOG.md
@@ -6,6 +6,11 @@
 
 Add `model` as an optional input argument to the `complete` method of `ChatCompletionsClient`.
 
+### Breaking Changes
+
+The field `input_tokens` was removed from class `EmbeddingsUsage`, as this was never defined in the
+REST API and the service never returned this value.
+
 ## 1.0.0b1 (2024-06-11)
 
 - Initial beta version